New_Commit

+++ server/controllers/bills/DELETE.js
@@ -7,7 +7,21 @@ export const DELETE_BILL = async (req, res) => {
 
 		// Find Bill
 		const bill = await Bills.findById(id);
+		const place = bill.place === "store" ? "مخزن" : "محل";
 		if (!bill) return res.status(400).json({ error: "حدث خطأ ولم يتم العثور علي الفاتورة" });
+
+		// Check If Products Count In Store OR Shop
+		const checkData = await bill.products.map(async ({ name, count }) => {
+			const company = await Products.findOne({ "products.name": name });
+			const product = company.products.find((item) => item.name === name);
+			return product?.count[bill.place] < count ? name : null;
+		});
+
+		const checkResult = (await Promise.all(checkData)).filter((item) => item);
+		if (checkResult.length)
+			return res
+				.status(200)
+				.json({ warn: `لا يتوفر الكمية المطلوبة من هذه المنتجات في ${place}: [${checkResult.join(" | ")}]` });
 
 		// Delete Bill
 		const deleteBill = await Bills.deleteOne({ _id: id });
@@ -24,16 +38,17 @@ export const DELETE_BILL = async (req, res) => {
 		});
 
 		const result = (await Promise.all(promises)).filter((item) => item);
-		const place = bill.place === "store" ? "مخزن" : "محل";
 		if (result.length)
 			return res.json({ warn: `حدث خطأ ولم يتم اضافه هذه المنتجات الي ${place} [${result.join(" | ")}]` });
 
 		// Send Transaction To Locker
 		const productsCost = bill.products.reduce((prev, cur) => prev + cur.price * cur.count, 0);
-		if (productsCost) await Locker.create({ name: `فاتورة ملغيه [${bill.client}]`, price: productsCost, method: "----" });
+		if (productsCost)
+			await Locker.create({ name: `فاتورة ملغيه [${bill.client}]`, price: productsCost, method: "----" });
 
 		res.status(200).json({ success: "لقد تم حذف الفاتورة بنجاح" });
 	} catch (error) {
+		console.log(error.stack);
 		res.status(404).json(`DELETE_BILL: ${error.message}`);
 	}
 };
